requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alibabacloud-credentials==1.0.7
alibabacloud-credentials-api==1.0.0
alibabacloud-darabonba-array==0.1.0
alibabacloud-darabonba-encode-util==0.0.2
alibabacloud-darabonba-map==0.0.1
alibabacloud-darabonba-signature-util==0.0.4
alibabacloud-darabonba-string==0.0.4
alibabacloud-endpoint-util==0.0.4
alibabacloud-gateway-pop==0.1.3
alibabacloud-gateway-spi==0.0.3
alibabacloud-kms20160120==2.4.0
alibabacloud-openapi-util==0.2.4
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.3
alibabacloud-tea-util==0.3.14
alibabacloud-tea-xml==0.0.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
av==16.1.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloup==3.0.8
colorama==0.4.6
cryptography==44.0.3
darabonba-core==1.0.5
decorator==5.2.1
fastapi==0.128.0
frozenlist==1.8.0
# glcontext==3.0.0  # Manim dependency
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
# isosurfaces==0.1.2  # Manim dependency
# manim==0.19.2  # Math animation engine
# manim-chemistry==0.5.1  # Manim chemistry extension
# ManimPango==0.6.1  # Manim text rendering
# mapbox_earcut==2.0.0  # Manim dependency
markdown-it-py==4.0.0
mdurl==0.1.2
# moderngl==5.12.0  # Manim OpenGL rendering
# moderngl-window==3.1.1  # Manim window management
multidict==6.7.1
nacos-sdk-python==2.0.9
networkx==3.6.1
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
# pycairo==1.29.0  # brew install pkg-config cairo  依赖这个系统库
pycparser==3.0
pycryptodome==3.23.0
pydantic[email]==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyglet==2.1.12
pyglm==2.8.3
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.1
scipy==1.17.0
screeninfo==0.8.1
six==1.17.0
skia-pathops==0.9.1
soupsieve==2.8.3
srt==3.5.3
starlette==0.50.0
svgelements==1.9.6
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
xmltodict==1.0.2
yarl==1.22.0
openpyxl==3.1.5
sqlalchemy==2.0.46
aiomysql==0.3.2
greenlet==3.3.1
passlib[argon2]==1.7.4
slowapi==0.1.9
redis==5.2.1
pydantic-settings>=2.12.0
python-jose==3.5.0
# akshare==1.18.30
yfinance>=0.2.50
openai==2.24.0
tushare~=1.4.24
fastmcp~=3.1.0
contextvars
# 核心 Web 框架与服务器
fastapi
uvicorn
watchfiles

# AI 与 大模型 SDK
google-genai
openai

# 配置与数据验证
pydantic
pydantic-settings
python-dotenv

# 工具与中間件
slowapi
requests
aiohttp
# 数据可视化
matplotlib